# Combat item during parsing (performance improvement)