--- a/ArcdpsLogManager/ArcdpsLogManager.csproj
+++ b/ArcdpsLogManager/ArcdpsLogManager.csproj
@@ -17,7 +17,7 @@
         Each new log data update causes a revision increase.
         See LogDataUpdater for the updates.
         -->
-        <Version>1.10.0.0</Version>
+        <Version>1.10.0.1</Version>
     </PropertyGroup>
     <ItemGroup>
         <PackageReference Include="Eto.Forms" Version="2.6.1" />