--- a/EVTCAnalytics/Parsed/ArcdpsBossIds.cs
+++ b/EVTCAnalytics/Parsed/ArcdpsBossIds.cs
@@ -1,0 +1,14 @@
+namespace GW2Scratch.EVTCAnalytics.Parsed;
+
+public static class ArcdpsBossIds
+{
+	/// <summary>
+	/// A world versus world log.
+	/// </summary>
+	public const int WorldVersusWorld = 1;
+	
+	/// <summary>
+	/// A full-instance log.
+	/// </summary>
+	public const int Map = 2;
+}