--- a/EVTCAnalytics/Parsing/AllowAllCombatItemFilters.cs
+++ b/EVTCAnalytics/Parsing/AllowAllCombatItemFilters.cs
@@ -1,0 +1,15 @@
+using GW2Scratch.EVTCAnalytics.Parsed.Enums;
+
+namespace GW2Scratch.EVTCAnalytics.Parsing;
+
+public class AllowAllCombatItemFilters : ICombatItemFilters {
+	public bool IsStateChangeRequired(StateChange stateChange)
+	{
+		return true;
+	}
+
+	public bool IsStateChangeRequired(byte stateChange)
+	{
+		return true;
+	}
+}