--- a/EVTCAnalytics/Parsing/IIdFilter.cs
+++ b/EVTCAnalytics/Parsing/IIdFilter.cs
@@ -1,0 +1,6 @@
+namespace GW2Scratch.EVTCAnalytics.Parsing;
+
+public interface IIdFilter
+{
+	bool IsKept(uint id);
+}