… ArcdpsLogManager/CHANGELOG.md
… EVTCAnalytics/Parsing/HashSetFilter.cs
… EVTCAnalytics/Parsing/ICombatItemFilters.cs
… EVTCAnalytics/Parsing/UnboundedArrayFilter.cs
… EVTCAnalytics/Processing/DefaultEncounterIdentifier.cs
… EVTCAnalytics/Processing/EncounterDataProvider.cs
… EVTCAnalytics/Processing/EncounterIdentifier.cs
… EVTCAnalytics/Processing/EncounterIdentifierBuilder.cs
… EVTCAnalytics/Processing/Encounters/IDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Modes/AgentHealthModeDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Modes/ConditionalModeDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Modes/ConstantModeDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Modes/EmboldenedDetectingModeDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Modes/FallbackModeDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Modes/GroupedSpawnModeDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Modes/IModeDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Modes/RemovedBuffStackRemainingTimeModeDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Modes/SkillPresentDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/AgentAliveCombatDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/AgentBelowHealthThresholdDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/AgentBuffGainedDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/AgentCombatExitDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/AgentDeadDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/AgentHealthResetDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/AgentKillingBlowDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/AgentOutsideOfCombatDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/BuffAppliedAfterSkillCastDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/BuffAppliedBelowHealthThresholdDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/ConstantResultDeterminer.cs
… EVTCAnalytics/Processing/Encounters/Results/EventFoundResultDeterminer.cs