Tested this pretty extensively, compared all of my 23k logs to see if there is any breakage, but probably wanna take one more look at it before merging just to make sure.

After this,
- All of my 23428 logs now take 9m9s to process
- 912 VG logs (39h of VG): 21.5s -> 15 seconds (1.43x faster)
- 578 instance logs (180 hours): 70 -> 36 seconds (1.94x faster)
- VG gathering, 4 threads: 11.5s -> 6.8s (1.7x faster)