@Sejsel 2023-08-22
When merging, keep in mind that `BuffExtensionEvent` was introduced in the meantime